requirements.txt
asgiref==3.5.2
autopep8==2.0.1
Django==4.1.4
djhtml==1.5.2
packaging==22.0
Pillow==9.4.0
pip-autoremove==0.10.0
pycodestyle==2.10.0
pytesseract==0.3.10
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
